two commit

+++ src/main/java/com/API_Finazas/app/rest/Model/Model.java
@@ -7,7 +7,7 @@ public class Model {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
     @Column
     private String producto;
 
@@ -29,10 +29,11 @@ public class Model {
     @Column
     private Integer user_id;
 
-    public long getId(){
+    public Integer getId() {
         return id;
     }
-    public void setId(long id){
+
+    public void setId(Integer id) {
         this.id = id;
     }
